# perf: optimize path parsing using strings.Count